+++ frontend/src/api/baseModel.ts
@@ -1,0 +1,8 @@
+export interface BasicListResult {
+  total: number;
+  next: string;
+  next_page_num: number;
+  previous: string;
+  previous_num: number;
+  results: any[];
+}