+++ frontend/src/api/projectInfo/project/project.ts
@@ -1,11 +1,11 @@
-import { ProjectInfoModel } from '../projectModel';
+import { ProjectInfoListResult, ProjectInfoModel } from '../projectModel';
 import { defHttp } from '/@/utils/http/axios';
 
 enum Api {
-  PROJECT_INFO = '/project_info/project_info',
+  PROJECT_INFO = '/project_info/project_info/',
 }
 
 // Get personal center-basic settings
 
 export const projectInfoList = () =>
-  defHttp.get<ProjectInfoModel>({ url: Api.PROJECT_INFO }, { joinTime: false });
+  defHttp.get<ProjectInfoListResult>({ url: Api.PROJECT_INFO }, { joinTime: false });