+++ frontend/types/axios.d.ts
@@ -33,7 +33,7 @@ export interface RetryRequest {
   waitTime: number;
 }
 export interface Result<T = any> {
-  code: number;
+  status: number;
   type: 'success' | 'error' | 'warning';
   message: string;
   result: T;